requirements.txt
attrs==25.3.0
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
contourpy==1.3.2
cycler==0.12.1
Flask==3.1.1
fonttools==4.58.4
gunicorn==23.0.0
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.3
numba==0.61.2
numpy==2.2.6
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.0
pillow==11.2.1
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
scikit-learn==1.7.0
scipy==1.15.3
selenium==4.33.0
shap==0.48.0
six==1.17.0
slicer==0.0.8
sniffio==1.3.1
sortedcontainers==2.4.0
threadpoolctl==3.6.0
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
